travel_booker/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
typer>=0.9.0
pydantic>=2.0.0
